adds ihave spam tests
